Added hilbert and z curve sorting

+++ math/vec3.cpp
@@ -6,7 +6,7 @@ vec3::vec3()
     setToZero();
 }
 
-vec3::vec3(float x, float y, float z)
+vec3::vec3(double x, double y, double z)
 {
     set(x,y,z);
 }
@@ -67,59 +67,59 @@ vec3 &vec3::operator/=(vec3 rhs) {
     return *this;
 }
 
-vec3 vec3::operator+(float scalar) {
+vec3 vec3::operator+(double scalar) {
     return vec3(m_vec[0] + scalar,
                 m_vec[1] + scalar,
                 m_vec[2] + scalar);
 }
 
-vec3 &vec3::operator+=(float scalar) {
+vec3 &vec3::operator+=(double scalar) {
     m_vec[0] += scalar;
     m_vec[1] += scalar;
     m_vec[2] += scalar;
     return *this;
 }
 
-vec3 vec3::operator-(float scalar) {
+vec3 vec3::operator-(double scalar) {
     return vec3(m_vec[0] - scalar,
                 m_vec[1] - scalar,
                 m_vec[2] - scalar);
 }
 
-vec3 &vec3::operator-=(float scalar) {
+vec3 &vec3::operator-=(double scalar) {
     m_vec[0] -= scalar;
     m_vec[1] -= scalar;
     m_vec[2] -= scalar;
     return *this;
 }
 
-vec3 vec3::operator*(float scalar) {
+vec3 vec3::operator*(double scalar) {
     return vec3(m_vec[0] * scalar,
                 m_vec[1] * scalar,
                 m_vec[2] * scalar);
 }
 
-vec3 &vec3::operator*=(float scalar) {
+vec3 &vec3::operator*=(double scalar) {
     m_vec[0] *= scalar;
     m_vec[1] *= scalar;
     m_vec[2] *= scalar;
     return *this;
 }
 
-vec3 vec3::operator/(float scalar) {
+vec3 vec3::operator/(double scalar) {
     return vec3(m_vec[0] / scalar,
                 m_vec[1] / scalar,
             m_vec[2] / scalar);
 }
 
-vec3 &vec3::operator/=(float scalar) {
+vec3 &vec3::operator/=(double scalar) {
     m_vec[0] /= scalar;
     m_vec[1] /= scalar;
     m_vec[2] /= scalar;
     return *this;
 }
 
-float vec3::dot(vec3 &rhs) {
+double vec3::dot(vec3 &rhs) {
     return (m_vec[0] * rhs.x() +
             m_vec[1] * rhs.y() +
             m_vec[2] * rhs.z());
@@ -131,12 +131,12 @@ vec3 vec3::cross(vec3 &rhs) {
                  m_vec[0] * rhs.y() - m_vec[1] * rhs.x());
 }
 
-float vec3::length() {
+double vec3::length() {
     return sqrt(lengthSquared());
 }
 
 void vec3::normalize() {
-    float myLength = length();
+    double myLength = length();
     if(myLength > 0) { // Don't divide by zero...
         m_vec[0] /= myLength;
         m_vec[1] /= myLength;
@@ -149,20 +149,20 @@ void vec3::setToZero()
     set(0,0,0);
 }
 
-void vec3::randomUniform(float min, float max) {
+void vec3::randomUniform(double min, double max) {
     m_vec[0] = min + Random::nextDouble()*(max - min);
     m_vec[1] = min + Random::nextDouble()*(max - min);
     m_vec[2] = min + Random::nextDouble()*(max - min);
 }
 
-void vec3::set(float x, float y, float z)
+void vec3::set(double x, double y, double z)
 {
     m_vec[0] = x;
     m_vec[1] = y;
     m_vec[2] = z;
 }
 
-void vec3::randomGaussian(float mean, float standardDeviation) {
+void vec3::randomGaussian(double mean, double standardDeviation) {
     m_vec[0] = Random::nextGaussian(mean, standardDeviation);
     m_vec[1] = Random::nextGaussian(mean, standardDeviation);
     m_vec[2] = Random::nextGaussian(mean, standardDeviation);
